refactor(editor): merge flow input-history into Cockpit's unified editor
Resolve the custom-editor slot conflict (notice.editorOwned warning) by
folding flow's ↑/↓ input-history into pi-cockpit's CockpitClaudeEditor so one
unified editor owns double-Escape clear, fullscreen markers, select-to-copy,
persistent history and the input route-target prefix — no more two extensions
racing for pi's single editor slot.

Cockpit:
- claude-editor.ts: CockpitClaudeEditor gains history browsing (↑/↓ walk, draft
  restore, position banner), addToHistory→persistent store, and the immutable
  route-target prefix render; factory accepts getEntries/record/getRouteTarget/
  onEditor. The redundant double-Escape inside the old HistoryEditor is dropped
  (DoubleEscapeGate already owns it).
- editor/input-history-store.ts: InputHistoryStore + workspaceStorageId moved
  from flow byte-identically (persisted history path unchanged).
- index.ts: installClaudeEditor is no longer gated on doubleEscape/fullscreen
  (history is a default feature); loads the store per session and sets the
  route target on the active editor directly (COCKPIT_INPUT_TARGET_EVENT still
  emitted for backward compat); store flushed on teardown.
- tui-i18n: History banner + unavailable messages (en/zh).

Flow:
- deletes src/tui/history-editor.ts + input-history.ts and their wiring
  (session start/shutdown, setInputRouteTarget, the COCKPIT_INPUT_TARGET_EVENT
  listener); no editor is installed by flow anymore.

Tests moved to cockpit: input-history-store.test.ts + unified-editor.test.ts
(history browsing, banner, route target, record, double-Escape coexistence).
flow cockpit-ownership drops the obsolete input-target assertions.

Verification: cockpit 702/702, new editor tests 15/15, flow cockpit-ownership
4/4, cockpit typecheck 0, RPC smoke success. No new flow failures vs baseline.
