+++ packages/pi-cockpit/src/tui-i18n.ts
@@ -291,6 +291,8 @@ export const COCKPIT_TUI_CATALOGS: TranslationCatalogs = {
 		"notice.editorOwned": "Cockpit double-Escape clear / fullscreen input unavailable: another extension owns the custom editor",
 		"notice.editorError": "Cockpit editor interaction error: {message}",
 		"notice.fullscreenDisabled": "Cockpit fullscreen input disabled: {message}",
+		"editor.historyBanner": "History {position}/{total}",
+		"editor.historyUnavailable": "Input history unavailable: {message}",
 		"notice.sidebarWidthSaveFailed": "Cockpit sidebar width kept for this session; save failed: {message}",
 		"notice.sidebarUnavailable": "Cockpit sidebar unavailable: {message}",
 		"notice.themeUnavailable": "theme \"{name}\" unavailable — {message}",
@@ -594,6 +596,8 @@ export const COCKPIT_TUI_CATALOGS: TranslationCatalogs = {
 		"notice.editorOwned": "Cockpit 双击 Esc 清空/全屏输入不可用：自定义编辑器由其他扩展占用",
 		"notice.editorError": "Cockpit 编辑器交互错误：{message}",
 		"notice.fullscreenDisabled": "Cockpit 全屏输入已禁用：{message}",
+		"editor.historyBanner": "历史 {position}/{total}",
+		"editor.historyUnavailable": "输入历史不可用：{message}",
 		"notice.sidebarWidthSaveFailed": "侧边栏宽度仅在本会话保留；保存失败：{message}",
 		"notice.sidebarUnavailable": "Cockpit 侧边栏不可用：{message}",
 		"notice.themeUnavailable": "主题 \"{name}\" 不可用 — {message}",
